Refactor overall architecture:
- add EIP55 address encoder
- convert address and hash fields to ByteArray and amount fields to BigInteger
- add TransactionSigner and TransactionBuilder
- remove contractAddress field from EthereumTransaction

+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/api/models/LastBlockHeight.kt
@@ -4,4 +4,4 @@ import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
 
 @Entity
-class LastBlockHeight(val height: Int, @PrimaryKey val id: String = "")
+class LastBlockHeight(val height: Long, @PrimaryKey val id: String = "")
